feat: 지수 데이터 등록 기능 추가

+++ src/main/java/com/codeit/findex/indexdata/dto/IndexDataDto.java
@@ -2,6 +2,7 @@ package com.codeit.findex.indexdata.dto;
 
 import com.codeit.findex.global.common.SourceType;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -16,17 +17,17 @@ public record IndexDataDto(
     @Schema(description = "출처 (사용자, Open API)", enumAsRef = true, example = "OPEN_API")
     SourceType sourceType,
     @Schema(description = "시가", example = "2800.25")
-    Double marketPrice,
+    BigDecimal marketPrice,
     @Schema(description = "종가", example = "2850.75")
-    Double closingPrice,
+    BigDecimal closingPrice,
     @Schema(description = "고가", example = "2870.5")
-    Double highPrice,
+    BigDecimal highPrice,
     @Schema(description = "저가", example = "2795.3")
-    Double lowPrice,
+    BigDecimal lowPrice,
     @Schema(description = "전일 대비 등락", example = "50.5")
-    Double versus,
+    BigDecimal versus,
     @Schema(description = "전일 대비 등락률", example = "1.8")
-    Double fluctuationRate,
+    BigDecimal fluctuationRate,
     @Schema(description = "거래량", example = "1250000")
     Long tradingQuantity,
     @Schema(description = "거래대금", example = "3500000000")
